requirements.txt
pydantic>=2.10.6
fastapi>=0.115.7
uvicorn>=0.18.3
ALMAFE-Lib>=0.0.16
websockets>=10.4
PyYAML>=6.0
keyboard>=0.13.5
openpyxl>=3.1.5
simple_pid>=2.0.1
pyinstrument>=5.0.1
python-can>=4.5.0
pyvisa>=1.14.1
nidaqmx>=1.0.2
nixnet>=0.3.2
pandas>=2.2.3
